fix output 0 task and add task 1

+++ 0x1E-search_algorithms/0-linear.c
@@ -1,9 +1,10 @@
 #include "search_algos.h"
 /**
- * 
- * 
- * 
- * 
+ * linear_search - searches for a value in an array of integers
+ * @array: pointer to point first position of array
+ * @size: size of array
+ * @value: value to find
+ * Return: position or -1
  */
 int linear_search(int *array, size_t size, int value)
 {
@@ -13,7 +14,7 @@ int linear_search(int *array, size_t size, int value)
 		return (-1);
 	for (i = 0; i < size; i++)
 	{
-		printf("Value checked array[%lu] = [%lu]\n", i, array[i]);
+		printf("Value checked array[%lu] = [%d]\n", i, (int)array[i]);
 		if (value == array[i])
 			return ((int)i);
 	}
